Respell the subframe count domain as Int64 for 32-bit platforms
Alternative to the saturating fix in #88, per review feedback: use Int64 where
the value genuinely needs 64 bits, rather than clamping.

At `.max100Days` a subframe count exceeds Int32.max for every frame rate at the
80- and 100-subframe bases (smallest: 23.976fps@80 = 16_588_800_000), so on a
32-bit platform — wasm32, watchOS armv7k/arm64_32 — the `Int` form trapped on
overflow. Because the bound is recomputed inside every wrapping add, that took
ALL arithmetic on a `.max100Days` timecode with it, however small the operands.

Widened, with no platform-conditional logic:
- internal: `FrameCount.subFrameCount`, `framesToSubFrames`, `subFramesToFrames`,
  `FrameCount.init(subFrameCount:base:)`, and the `sfcNew` locals (which infer).
- public: `TimecodeFrameRate.maxTotalSubFrames(in:base:)`,
  `maxSubFrameCountExpressible(in:base:)`, `Timecode.maxSubFrameCountExpressible`.

Deliberately NOT widened: `maxTotalFrames`, which peaks at 1_036_800_000
(120fps @ 100 days) and fits a 32-bit Int; and the frames/subFrames components,
which are bounded by it. Only the COUNT needs 64 bits.

One narrowing remains, at `Timecode.rationalValue`: `Fraction` is Int-based, so
a timecode beyond ~Int32.max subframes has no representable rational value on a
32-bit platform. That is a pre-existing limit of `Fraction`, not of the count,
and is commented at the site.

Unlike the saturating approach this makes `.max100Days` genuinely usable on
32-bit rather than merely non-trapping.

Full suite passes: 506 tests in 55 suites.

+++ Sources/SwiftTimecodeCore/Timecode/FrameCount/FrameCount.swift
@@ -105,7 +105,7 @@ extension Timecode {
 
 extension Timecode.FrameCount {
     init(
-        subFrameCount: Int,
+        subFrameCount: Int64,
         base: Timecode.SubFramesBase
     ) {
         let converted = Timecode.subFramesToFrames(
@@ -225,7 +225,7 @@ extension Timecode.FrameCount {
     public func multiplying(by factor: Double) -> Self {
         let lhsTotalSubFrames = subFrameCount
 
-        let resultSubFrameCount = Int(Double(lhsTotalSubFrames) * factor)
+        let resultSubFrameCount = Int64(Double(lhsTotalSubFrames) * factor)
 
         let newFrames = Timecode.subFramesToFrames(
             resultSubFrameCount,
@@ -241,7 +241,7 @@ extension Timecode.FrameCount {
     public func dividing(by divisor: Double) -> Self {
         let lhsTotalSubFrames = subFrameCount
 
-        let resultSubFrameCount = Int(Double(lhsTotalSubFrames) / divisor)
+        let resultSubFrameCount = Int64(Double(lhsTotalSubFrames) / divisor)
 
         let newFrames = Timecode.subFramesToFrames(
             resultSubFrameCount,
@@ -268,7 +268,7 @@ extension Timecode.FrameCount {
 }
 
 extension Timecode.FrameCount {
-    var subFrameCount: Int {
+    var subFrameCount: Int64 {
         Timecode.framesToSubFrames(
             frames: wholeFrames,
             subFrames: subFrames,
@@ -285,15 +285,19 @@ extension Timecode {
         frames: Int,
         subFrames: Int,
         base: SubFramesBase
-    ) -> Int {
-        (frames * base.rawValue) + subFrames
+    ) -> Int64 {
+        // Int64 because a 100-day timecode's subframe count exceeds Int32.max
+        // for every frame rate, and `Int` is 32-bit on wasm32 / watchOS armv7k.
+        (Int64(frames) * Int64(base.rawValue)) + Int64(subFrames)
     }
 
     /// Internal utility
-    static func subFramesToFrames(_ subFrames: Int, base: SubFramesBase) -> (frames: Int, subFrames: Int) {
-        let outSubFrames = subFrames % base.rawValue
-        let outFrames = (subFrames - outSubFrames) / base.rawValue
+    static func subFramesToFrames(_ subFrames: Int64, base: SubFramesBase) -> (frames: Int, subFrames: Int) {
+        // The COUNT needs 64 bits; the resulting frames/subFrames do not —
+        // max total frames is ~1.04e9 even at 120 fps over 100 days.
+        let outSubFrames = subFrames % Int64(base.rawValue)
+        let outFrames = (subFrames - outSubFrames) / Int64(base.rawValue)
 
-        return (frames: outFrames, subFrames: outSubFrames)
+        return (frames: Int(outFrames), subFrames: Int(outSubFrames))
     }
 }
